+++ src/main/java/com/debateApp/Main/services/MessageService.java
@@ -1,5 +1,8 @@
 package com.debateApp.Main.services;
 
+import org.apache.tomcat.util.http.parser.Authorization;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.debateApp.Main.repositories.*;
@@ -7,6 +10,7 @@ import com.debateApp.Main.dto.MessageResponseDTO;
 import com.debateApp.Main.dto.AddMessageDTO;
 import com.debateApp.Main.entities.Messages;
 import com.debateApp.Main.entities.Stance;
+import com.debateApp.Main.entities.Groups;
 
 import lombok.RequiredArgsConstructor;
 
@@ -19,6 +23,20 @@ public class MessageService {
     private final UserRepository userRepository;
 
     public MessageResponseDTO addMessage(AddMessageDTO dto) {
+
+        Long userId = (Long) SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+        Groups group = groupRepository.findById(dto.getGroupId())
+                .orElseThrow(() -> new RuntimeException("Group not found, id : " + dto.getGroupId()));
+
+        boolean isMember = group.getMembers().stream()
+                .anyMatch(member -> member.getId().equals(userId));
+
+        if (!isMember) {
+            throw new AccessDeniedException("Only members can message in the group.");
+        }
 
         Messages message = new Messages();
 
@@ -33,14 +51,13 @@ public class MessageService {
         message.setGroup(groupRepository.findById(dto.getGroupId())
                 .orElseThrow(() -> new RuntimeException("Group not found, id : " + dto.getGroupId())));
 
-        message.setAuthor(userRepository.findById(dto.getUserId())
-                .orElseThrow(() -> new RuntimeException("User not found, id : " + dto.getUserId())));
+        message.setAuthor(userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("User not found, id : " + userId)));
 
-        // TODO
-        // User JWT to authenticate the user.
-        if (dto.getParentId() != null)
+        if (dto.getParentId() != null) {
             message.setParent(messageRepository.findById(dto.getParentId())
                     .orElseThrow(() -> new RuntimeException("Parent message not found, id : " + dto.getParentId())));
+        }
 
         messageRepository.save(message);
 
@@ -68,6 +85,17 @@ public class MessageService {
     }
 
     public void deleteMessage(Long id) {
+        Messages message = messageRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Group not found, id : " + id));
+
+        Long userId = (Long)SecurityContextHolder.getContext()
+            .getAuthentication()
+            .getPrincipal();
+
+        if(!message.getAuthor().getId().equals(userId)){
+            throw new AccessDeniedException("Only the author of this message can delete it!");
+        }
+
         messageRepository.deleteById(id);
     }
 }
